Add files via upload
This has integrated combat function. 
Added comments to a lot of the program to look more thourough when editing.

Combat function returns a tuple of "true" when enemy health is less than 0, and a value of the player health from combat (Damage)
Only input is "4" for attack.
Will have to redo the original function that allows a player to run while keeping the room undiscovered.